+++ Localization/deDE.lua
@@ -1,4 +1,82 @@
 local L = LibStub('AceLocale-3.0'):NewLocale('PetBattleScripts', 'deDE')
 if not L then return end
 
-@localization(locale="deDE")@
+--[[ L["ADDON_NAME"] = "Pet Battle Scripts" --]]
+L["DATABASE_UPDATE_BASE_TO_FIRSTENEMY_NOTIFICATION"] = "Es wurde festgestellt, dass du zuvor eine modifizierte Version von tdBattlePetScript verwendet hast und einige Skripte des modifizierten Basis-Selektors in den Erster Gegner-Selektor migriert hast."
+L["DATABASE_UPDATED_TO"] = "Aktualisiert auf Version: "
+L["DEFAULT_NEW_SCRIPT_NAME"] = "Neues Skript"
+L["DIRECTOR_TEST_NEXT_ACTION"] = "Nächste Aktion"
+L["EDITOR_CREATE_SCRIPT"] = "Skript erstellen"
+L["EDITOR_EDIT_SCRIPT"] = "Skript bearbeiten"
+L["IN_BATTLE_DEBUGGING_SCRIPT"] = "Skript debuggen"
+--[[ L["IN_BATTLE_EXECUTE"] = "Autobattle" --]]
+L["IN_BATTLE_NO_SCRIPT"] = "Kein Skript"
+L["IN_BATTLE_SELECT_SCRIPT"] = "Skript auswählen"
+L["OPTION_AUTO_SELECT_SCRIPT_BY_ORDER"] = "Skripte automatisch auf Basis der Skript-Selektor-Prioritäten auswählen"
+L["OPTION_AUTO_SELECT_SCRIPT_ONLY_ONE"] = "Automatisch nur ein Skript auswählen"
+L["OPTION_AUTOBUTTON_HOTKEY"] = "Hotkey für die Auto-Schaltfläche"
+L["OPTION_AUTOBUTTON_HOTKEY_SECONDARY"] = "(alternativ)"
+L["OPTION_EDITOR_FONT_FACE"] = "Schriftart"
+L["OPTION_EDITOR_FONT_SIZE"] = "Schriftgröße"
+L["OPTION_HIDE_MINIMAP"] = "Minimap-Symbol verstecken"
+L["OPTION_HIDE_SELECTOR_NO_SCRIPT"] = "Skript-Selektor nicht anzeigen, wenn keine Skripte verfügbar sind"
+L["OPTION_LOCK_SCRIPT_SELECTOR"] = "Skript-Selektor sperren (kann nicht verschoben werden)"
+L["OPTION_NO_WAIT_DELETE_SCRIPT"] = "Nicht auf das Löschen eines Skripts warten"
+L["OPTION_NOTIFY_BUTTON_ACTIVE"] = "Spiele Ton wenn \"Autobattle\" Schaltfläche aktiv wird"
+L["OPTION_NOTIFY_BUTTON_ACTIVE_SOUND"] = "Ton"
+L["OPTION_RESET_FRAMES"] = "Größe und Position von allen Fenstern zurücksetzen"
+L["OPTION_SCRIPTSELECTOR_NOTES"] = "Hier kannst du die Verfügbarkeit und Priorität der Skript-Selektoren einstellen."
+L["OPTION_SETTINGS_HIDE_MINIMAP_TOOLTIP"] = "Um diese Einstellung zu ändern, muss die Benutzeroberfläche neu geladen werden. Willst du fortfahren?"
+L["OPTION_TEST_BREAK"] = "Entwickler: Die 'test' Aktion unterbricht die Ausführung"
+L["REMATCH_NOTE_SCRIPT_EXPORT_ADD_TO_NOTE_MENU_ITEM"] = "Füge Skript an Notiz dran"
+L["REMATCH_NOTE_SCRIPT_IMPORT_FAIL"] = [=[Mindestens ein Import von Skripten aus Rematch Team Notizen ist fehlgeschlagen:
+%s]=]
+L["REMATCH_NOTE_SCRIPT_IMPORT_FAIL_EXIST_DIFFERENT"] = "Ein Skript existiert bereits, aber es ist unterschiedlich. Lösche Notiz oder Skript."
+L["REMATCH_NOTE_SCRIPT_IMPORT_FAIL_LINE"] = "- Team \"%s\": %s"
+L["REMATCH4_DEPRECATED"] = [=[Rematch 4 ist alt und die Unterstützung seitens des Pet Battle Script addons wird Mitte 2025 beendet. Bitte update zu Rematch 5.
+
+Zusätzlich, sag uns bitte auf Curseforge oder GitHub Bescheid, damit wir zählen können, wieviele Nutzer noch Rematch 4 benutzen.]=]
+L["SCRIPT_EDITOR_AUTOFORMAT_SCRIPT"] = "Skript formatieren"
+L["SCRIPT_EDITOR_DELETE_SCRIPT_CONFIRMATION"] = "Bist du sicher, dass du das Skript |cffffd000[%s - %s]|r |cffff0000löschen|r willst?"
+L["SCRIPT_EDITOR_FOUND_ERROR"] = "Fehler gefunden"
+L["SCRIPT_EDITOR_LABEL_TOGGLE_EXTRA"] = "Plugin-Daten-Editor öffnen/schließen"
+L["SCRIPT_EDITOR_NAME_TITLE"] = "Skriptname"
+L["SCRIPT_EDITOR_RUN_BUTTON"] = "Ausführen"
+L["SCRIPT_EDITOR_SAVE_SUCCESS"] = "Erfolgreich gespeichert"
+L["SCRIPT_EDITOR_TEXTAREA_TITLE"] = "Skript"
+L["SCRIPT_EDITOR_TITLE"] = "Skript-Editor"
+L["SCRIPT_MANAGER_TITLE"] = "Skript-Manager"
+L["SCRIPT_MANAGER_TOGGLE"] = "Skript-Manager öffnen/schließen"
+L["SCRIPT_SELECTOR_TITLE"] = "Skript-Selektor"
+L["SCRIPT_SELECTOR_TOGGLE"] = "Skript-Selektor öffnen/schließen"
+L["SELECTOR_ALLINONE_NOTES"] = "Dieses Skript kann in allen Haustierkämpfen verwendet werden."
+L["SELECTOR_ALLINONE_TITLE"] = "Alle in einem"
+L["SELECTOR_BASE_ALLY"] = "Verbündeter"
+L["SELECTOR_BASE_ENEMY"] = "Feind"
+L["SELECTOR_BASE_NOTES"] = "Dieser Skript-Selektor gilt für Verbündeten und Feind."
+L["SELECTOR_BASE_TITLE"] = "Basis"
+L["SELECTOR_FIRSTENEMY_NOTES"] = "Dieser Skript-Selektor gilt für den ersten Gegner des Kampfes."
+L["SELECTOR_FIRSTENEMY_TITLE"] = "Erster Gegner"
+L["SELECTOR_REMATCH_4_TO_5_UPDATE_NOTE"] = [=[Von Rematch 4 zu Rematch 5 konvertiert. Bitte überprüfe, ob alle Deine Skripte weiterhin korrekt mit Teams verbunden sind.\n
+Wenn die Konvertierung fehlgeschlagen ist kannst du ein Backup von wow/WTF/Account/<account>/SavedVariables/tdBattlePetScript.lua wiederherstellen, oder diese Datei öffnen, nach "Rematch" suchen und es mit "Rematch5" ersetzen, dann "Rematch4" suchen und mit "Rematch" ersetzen. Danach installiere wieder Rematch 4 und berichte uns das Problem unter https://github.com/axc450/pbs/issues/new mit der SavedVariables Datei von Rematch und diesem Addon.]=]
+L["SELECTOR_REMATCH_4_TO_5_UPDATE_ORPHAN"] = [=[Habe Skript "%s" gefunden das mit der nicht-existenten Rematch Team ID "%s" verbunden ist.\n
+Das kann durch ein Problem während dem Datenbankupdate passiert sein, oder bereits vorher kaputt gewesen sein. Wenn der Fehler bei vielen Teams auftaucht, dann berichte uns bitte von dem Problem. Ansonsten entferne die kaputten Teams mit dem Skript-Manager und füg sie manuell wieder dem korrekten Team zu.]=]
+L["SELECTOR_REMATCH_CANT_FORMAT_TOOLTIP_REMATCH_NOT_LOADED"] = "Kann keine Informationen zeigen: Das Rematch-Addon ist nicht geladen."
+L["SELECTOR_REMATCH_NO_TEAM_FOR_SCRIPT"] = "Kein Team passt zu diesem Skript"
+L["SELECTOR_REMATCH_NOTES"] = "Dieser Skript-Selektor gilt basiered auf dem aktuellen Rematch Team."
+--[[ L["SELECTOR_REMATCH_TEAM_FORMAT"] = "Team: %s" --]]
+--[[ L["SELECTOR_REMATCH_TITLE"] = "Rematch" --]]
+L["SHARE_EXPORT_SCRIPT"] = "Exportieren"
+L["SHARE_IMPORT_CHOOSE_KEY"] = "Wert auswählen..."
+L["SHARE_IMPORT_CHOOSE_SELECTOR"] = "Skript-Selektor auswählen..."
+L["SHARE_IMPORT_LABEL_ALREADY_EXISTS_CHECKBOX"] = "Überschreiben und Importieren fortsetzen"
+L["SHARE_IMPORT_LABEL_ALREADY_EXISTS_WARNING"] = "Das Skript existiert bereits mit diesen Einstellungen. Wenn du mit der Importierung fortfährst, wird das aktuelle Skript überschrieben."
+L["SHARE_IMPORT_LABEL_HAS_EXTRA"] = "Dieser Import-String importiert zusätzlich zum Skript-Code Daten die vom Plugin abhängen. Meistens sind das Informationen über das zugehörige Team."
+L["SHARE_IMPORT_PLUGIN_NOT_ENABLED"] = "Importieren nicht möglich: Plugin ist nicht aktiviert."
+L["SHARE_IMPORT_REINPUT_TEXT"] = "Erneut bearbeiten"
+L["SHARE_IMPORT_SCRIPT"] = "Importieren"
+L["SHARE_IMPORT_SCRIPT_EXISTS"] = "Skript bereits vorhanden"
+L["SHARE_IMPORT_SCRIPT_NOT_IMPORT_STRING_WARNING"] = "Es wird empfohlen, den Import mit einem gültigen Share-String anstelle eines manuell eingegebenen Skripts durchzuführen. Dennoch fortfahren?"
+L["SHARE_IMPORT_SCRIPT_WELCOME"] = "Füge Share-String oder Skript in das Eingabefeld ein."
+L["SHARE_IMPORT_STRING_INCOMPLETE"] = "Die Daten des Share-Strings sind zwar unvollständig, aber können trotzdem importiert werden."
+L["TOOLTIP_CREATE_OR_DEBUG_SCRIPT"] = "Ein Skript erstellen oder debuggen"