+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Other
 
+- Localization is no longer maintained on Curseforge but within the GitHub repository.
 - Bumped TOC for patch 11.1.7
 
 ## v1.11.4